Move form store logic into its own component.
